document and organize
